ml-service/requirements.txt
fastapi
uvicorn
lightgbm
pandas
numpy
scikit-learn
psycopg2-binary
python-dotenv
